Comments and documentation

+++ src/main/java/com/dashroshan/Controller.java
@@ -100,6 +100,7 @@ public class Controller {
     @FXML
     private ImageView saveButton;
 
+    // Enable or disable the output setting controls
     private void switchControls(boolean disabled) {
         outputResolutionHeight.setDisable(disabled);
         outputResolutionWidth.setDisable(disabled);
@@ -117,19 +118,25 @@ public class Controller {
         saveButton.setCursor(disabled ? Cursor.DEFAULT : Cursor.HAND);
     }
 
+    // Initialize the video thumbnail, info, and default output settings
     private void setThumbnailAndInfo() throws IOException {
+        // Load FFMPEG
         FFprobe ffprobe = new FFprobe("ffprobe.exe");
         FFmpeg ffmpeg = new FFmpeg("ffmpeg.exe");
         FFmpegExecutor executor = new FFmpegExecutor(ffmpeg, ffprobe);
 
+        // Get info about the input video
         FFmpegProbeResult in = ffprobe.probe(inputFilePath);
         FFmpegFormat format = in.getFormat();
         FFmpegStream stream = in.getStreams().get(0);
 
+        // Set input video info in the preview area
         inputFileName.setText(inputFilePath.substring(inputFilePath.lastIndexOf('\\') + 1));
         inputCodec.setText(stream.codec_name.toUpperCase() + " Codec");
         inputDuration.setText(String.format("%.2fs", format.duration));
         inputResolution.setText(String.format("%d x %dpx", stream.width, stream.height));
+
+        // Set default output settings same as that of input
         originalRatio = (double) stream.width / (double) stream.height;
         originalHeight = stream.height;
         originalWidth = stream.width;
@@ -150,26 +157,50 @@ public class Controller {
                 .select(inputFilePath.substring(inputFilePath.lastIndexOf('.') + 1).toUpperCase());
         resolutionRatio.getSelectionModel().select("Variable");
 
+        // Ready FFMPEG to extract the thumbnail
         String outputFilePath = inputFilePath.substring(0, inputFilePath.lastIndexOf('\\'));
-
         FFmpegBuilder builderThumb = new FFmpegBuilder()
                 .setInput(in)
                 .addOutput(outputFilePath + "thumb.jpg").addExtraArgs("-vframes", "1")
                 .done();
-        executor.createJob(builderThumb).run();
 
-        File thumbImg = new File(outputFilePath + "thumb.jpg");
-        Image fxImage = new Image(thumbImg.toURI().toString());
-        videoThumbnail.setImage(fxImage);
-        thumbImg.delete();
+        // To avoid blocking the JavaFX UI thread, run the thumbnail extraction using
+        // FFMPEG in another thread
+        FFmpegJob job = executor.createJob(builderThumb, new ProgressListener() {
+            @Override
+            public void progress(Progress progress) {
+                // When thumbnail extracted update UI in runLater as JavaFX UI thread might not
+                // be free at this instant
+                Platform.runLater(new Runnable() {
+                    @Override
+                    public void run() {
+                        if (progress.status.toString().equals("end")) {
+                            // The thumbnail is saved in same location as input video
+                            // Load it to the imageview and then delete
+                            File thumbImg = new File(outputFilePath + "thumb.jpg");
+                            Image fxImage = new Image(thumbImg.toURI().toString());
+                            videoThumbnail.setImage(fxImage);
+                            thumbImg.delete();
 
-        switchControls(false);
+                            // Enable the output setting controls
+                            switchControls(false);
+                        }
+                    }
+                });
+            }
+        });
+
+        // Set thread as daemon to avoid conversion continuing if app is closed
+        // Then start the thread
+        Thread processingThread = new Thread(job);
+        processingThread.setDaemon(true);
+        processingThread.start();
     }
 
+    // Called when user clicks the Save button
     @FXML
     protected void convertAndSave() throws IOException {
-        // Create a file chooser window for the user to select output file name and
-        // location
+        // Create a file chooser window to select output file name and location
         FileChooser file_chooser = new FileChooser();
 
         // Set the video format from Output format selection box in the file chooser
@@ -183,8 +214,10 @@ public class Controller {
         if (file == null)
             return;
 
+        // Disable all output setting controls
         switchControls(true);
 
+        // Calculate the height and width based on the ratio dropdown box
         String ratio = resolutionRatio.getValue();
         int opHeight, opWidth;
         if (ratio == "Fixed height") {
@@ -198,11 +231,14 @@ public class Controller {
             opHeight = Integer.parseInt(outputResolutionHeight.getText());
         }
 
+        // Most video codecs dont allow odd height or width in pixels
+        // So make both even if not already
         if (opWidth % 2 == 1)
             opWidth++;
         if (opHeight % 2 == 1)
             opHeight++;
 
+        // Get output video settings
         double opQuality = outputQuality.getValue() / 100.0;
         long startH = Long.parseLong(outputDurationStartH.getText());
         long startM = Long.parseLong(outputDurationStartM.getText());
@@ -213,14 +249,16 @@ public class Controller {
         long endS = Long.parseLong(outputDurationEndS.getText());
         long end = endH * 3600 + endM * 60 + endS;
 
+        // Get output file path
         String outputFilePath = file.getAbsolutePath();
 
+        // Load FFMPEG
         FFprobe ffprobe = new FFprobe("ffprobe.exe");
         FFmpeg ffmpeg = new FFmpeg("ffmpeg.exe");
 
+        // Load the video into FFMPEG along with the output video settings
         FFmpegExecutor executor = new FFmpegExecutor(ffmpeg, ffprobe);
         FFmpegProbeResult in = ffprobe.probe(inputFilePath);
-
         FFmpegBuilder builder = new FFmpegBuilder()
                 .setInput(in)
                 .addOutput(outputFilePath)
@@ -232,15 +270,19 @@ public class Controller {
                 .setDuration(end - start, TimeUnit.SECONDS)
                 .done();
 
+        // To avoid blocking the JavaFX UI thread, run the conversion in another thread
         FFmpegJob job = executor.createJob(builder, new ProgressListener() {
             final double inputDuration_ns = (end - start) * TimeUnit.SECONDS.toNanos(1);
 
+            // Gets called for every 1% change in conversion progress
             @Override
             public void progress(Progress progress) {
+                // Update progressbar in runLater as UI thread might not be free this instant
                 Platform.runLater(new Runnable() {
                     @Override
                     public void run() {
                         if (progress.status.toString().equals("end")) {
+                            // At end, reset progressbar to 0 and enable the output setting controls again
                             progressBar.setProgress(0);
                             switchControls(false);
                         } else
@@ -250,16 +292,21 @@ public class Controller {
             }
         });
 
+        // Set thread as daemon to avoid conversion continuing if app is closed
+        // Then start the thread
         Thread processingThread = new Thread(job);
         processingThread.setDaemon(true);
         processingThread.start();
     }
 
+    // Called when users drags some file over the drag and drop area but hadnot
+    // dropped it yet
     @FXML
     protected void dragAndDropDragOver(DragEvent event) {
         Dragboard db = event.getDragboard();
 
         if (event.getGestureSource() != dragAndDrop && db.hasFiles()) {
+            // If dragged file is in a supported video format allow the user to drop it
             String fileName = db.getFiles().get(0).getName().toLowerCase();
             if (Utility.isValidVideoFile(fileName))
                 event.acceptTransferModes(TransferMode.COPY_OR_MOVE);
@@ -267,6 +314,7 @@ public class Controller {
             event.consume();
     }
 
+    // Called when user drops a file over the drag and drop area
     @FXML
     protected void dragAndDropDragDropped(DragEvent event) throws IOException {
         Dragboard db = event.getDragboard();
@@ -274,27 +322,34 @@ public class Controller {
         if (db.hasFiles()) {
             inputFilePath = db.getFiles().get(0).toString();
             success = true;
+
+            // Show input video thumbnail and info
+            setThumbnailAndInfo();
         }
-        setThumbnailAndInfo();
         event.setDropCompleted(success);
         event.consume();
     }
 
+    // Handles click event for the drag and drop area to select the input file
     @FXML
     protected void dragAndDropClicked() throws IOException {
+        // Opens a filechooser with available video extension options
         FileChooser file_chooser = new FileChooser();
         file_chooser.getExtensionFilters().addAll(
                 new FileChooser.ExtensionFilter(
                         "Video file", Utility.extensionsPattern));
-
         File file = file_chooser.showOpenDialog(videoThumbnail.getScene().getWindow());
 
+        // Checks if user clicked cancel or selected an unsupported file type
         if (file != null && Utility.isValidVideoFile(file.getName())) {
             inputFilePath = file.getAbsolutePath();
+
+            // Show input video thumbnail and info
             setThumbnailAndInfo();
         }
     }
 
+    // Controls output resolution width and height based on the selected ratio type
     @FXML
     protected void ratioChanged(ActionEvent action) {
         String ratio = resolutionRatio.getValue();
@@ -316,18 +371,23 @@ public class Controller {
         }
     }
 
+    // Shows help info alert
+    // Help info texts for given title is stored in the Info class
     private void showHelp(String title) {
         Alert alert = new Alert(AlertType.NONE, Info.messages.get(title), ButtonType.OK);
         alert.getDialogPane().setMinHeight(Region.USE_PREF_SIZE);
 
+        // Set favicon for the alert window
         Stage stage = (Stage) alert.getDialogPane().getScene().getWindow();
         Image favicon = new Image(App.class.getResource("icon.png").toExternalForm());
         stage.getIcons().add(favicon);
 
+        // Show alert
         alert.setTitle(title);
         alert.show();
     }
 
+    // Help info alerts for all output settings available
     @FXML
     protected void helpResolution() {
         showHelp("Resolution");
